Convert list of annual raster stack to data.table

Convert list of annual raster stack to data.table

annualStackToDTx1000(x, whNotNA, ...)

# S3 method for RasterLayer
annualStackToDTx1000(x, whNotNA, ...)

# S3 method for list
annualStackToDTx1000(x, whNotNA, ...)

# S3 method for RasterStack
annualStackToDTx1000(x, whNotNA, ...)

Arguments

x
RasterStack or list of rasters to convert to data.table and multiply by 1000 to save space

whNotNA
Pixel indexes that should go through this process (i.e. not NA)

...
Not currently used

Value

data.table of the raster stack or the list

Examples

library(raster)

#> Loading required package: sp

r1 <- raster(extent(0,10,0,10), vals = 1:100)
r2 <- raster(extent(0,10,0,10), vals = 100:1)
r3 <- raster(extent(0,10,0,10), vals = 200:101)
r4 <- raster(extent(0,10,0,10), vals = 300:201)

# list of Rasters
lRast <- list(r1, r2, r3)
lRast[[1]][5] <- NA
whNotNA <- setdiff(1:ncell(r1), 5)

# unnamed -- should error
try(out1 <- annualStackToDTx1000(lRast, whNotNA))

#> Error in annualStackToDTx1000.list(lRast, whNotNA) : 
#>   x must be a named list (or stack)

# named
names(lRast) <- c("OneToHun", "HunToOne", "TwoHunToOneHun")
out1 <- annualStackToDTx1000(lRast, whNotNA)

#> Layer layer converted to data.table

#> Layer layer converted to data.table

#> Layer layer converted to data.table

# RasterStack
out2 <- annualStackToDTx1000(raster::stack(lRast), whNotNA)

#> Layer OneToHun converted to data.table

#> Layer HunToOne converted to data.table

#> Layer TwoHunToOneHun converted to data.table

# List of RasterStacks
s1 <- raster::stack(r1, r2)
names(s1) <- names(lRast)[1:2]
s2 <- raster::stack(r4, r3)
names(s2) <- c(names(lRast)[3], "ThreeHunToTwoHun")
out3 <- annualStackToDTx1000(list(s1 = s1, s2 = s2), whNotNA) ## named list required

#> Layer OneToHun converted to data.table

#> Layer HunToOne converted to data.table

#> Layer TwoHunToOneHun converted to data.table

#> Layer ThreeHunToTwoHun converted to data.table

# With duplicated names -- to remove duplicates;
#  actually, this doesn't make sense: RasterStack can't have duplicated names
names(lRast) <- c("OneToHun", "OneToHun", "TwoHunToOneHun")
out4 <- annualStackToDTx1000(raster::stack(lRast), whNotNA)

#> Layer OneToHun.1 converted to data.table

#> Layer OneToHun.2 converted to data.table

#> Layer TwoHunToOneHun converted to data.table