Wrapper around DEoptim call

Does the multiple cluster connections. This will only work if ssh keys are correctly made between machines (if using multiple machines).

runDEoptim(
  landscape,
  annualDTx1000,
  nonAnnualDTx1000,
  fireBufferedListDT,
  historicalFires,
  itermax,
  initialpop,
  NP,
  trace,
  strategy,
  cores,
  logPath,
  cachePath,
  iterStep = 25,
  lower,
  upper,
  formula,
  objFunCoresInternal,
  covMinMax = covMinMax,
  tests,
  maxFireSpread,
  Nreps,
  .verbose,
  visualizeDEoptim
)

DEoptimIterative(
  itermax,
  lower,
  upper,
  control,
  formula,
  covMinMax,
  tests,
  objFunCoresInternal,
  maxFireSpread,
  Nreps,
  visualizeDEoptim,
  cachePath,
  iterStep = 25,
  .verbose
)

Arguments

landscape
A RasterLayer which has the correct metadata associated with the pixelID and cells of other objects in this function call

annualDTx1000
A list of data.table objects. Each list element will be from 1 year, and it must be the same length as fireBufferedListDT and historicalFires

nonAnnualDTx1000
A list of data.table objects. Each list element must be named with a concatenated sequence of names from names(annualDTx1000), e.g., 1991_1992_1993. It should contain all the years in names(annualDTx1000).

fireBufferedListDT
A list of data.table objects. It must be same length as annualDTx1000, with same names. Each element is a data.table with columns: buff...TODO: INCOMPLETE

historicalFires
DESCRIPTION NEEDED

itermax
Passed to DEoptim.control

initialpop
DESCRIPTION NEEDED

NP
DESCRIPTION NEEDED

trace
Passed to DEoptim.control

strategy
Passed to DEoptim.control

cores
A numeric (for running on localhost only) or a character vector of machine names (including possibly "localhost"), where the length of the vector indicates how many cores should be used on that machine.

logPath
A character string indicating what file to write logs to. This dirname(logPath) must exist on each machine, though the function will make sure it does internally.

cachePath
The cachePath to store cache in. Should likely be cachePath(sim)

iterStep
Integer. Must be less than itermax. This will cause DEoptim to run the itermax iterations in ceiling(itermax / iterStep) steps. At the end of each step, this function will plot, optionally, the parameter histograms (if visualizeDEoptim is TRUE)

lower
Passed to DEoptim

upper
Passed to DEoptim

formula
Passed to DEoptim

objFunCoresInternal
DESCRIPTION NEEDED

covMinMax
Passed to fireSenseUtils::.objfun

tests
Passed to fireSenseUtils::.objfun

maxFireSpread
Passed to fireSenseUtils::.objfun

Nreps
Passed to fireSenseUtils::.objfun

.verbose
Passed to fireSenseUtils::.objfun

visualizeDEoptim
Logical. If TRUE, then histograms will be made of DEoptim outputs

control
DESCRIPTION NEEDED

Value

DESCRIPTION NEEDED